feat: errorModal 추가
